Update
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@8090 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/filed/xattr.c
@@ -202,7 +202,7 @@ static uint32_t serialize_xattr_stream(JCR *jcr, uint32_t expected_serialize_len
 static bool generic_xattr_build_streams(JCR *jcr, FF_PKT *ff_pkt, int stream)
 {
    int count = 0;
-   ssize_t xattr_list_len,
+   int32_t xattr_list_len,
            xattr_value_len,
            expected_serialize_len = 0,
            serialize_len = 0;
